moved tests to the data layer; changed valid course bounds to accurately reflect allowed course numbers
