+++ cli/data/data_test.go
@@ -1,0 +1,45 @@
+package data
+
+import (
+	"testing"
+)
+
+func TestValidName(t *testing.T) {
+	_, err := GetCourseByName("Data Structures")
+
+	if err != nil {
+		t.Errorf("error when getting valid course")
+	}
+
+	_, err = GetCourseByName("")
+
+	if err == nil {
+		t.Errorf("should not be allowed to get empty course")
+	}
+}
+
+func TestValidNumber(t *testing.T) {
+	_, err := GetCourseByNum("CS", 225)
+
+	if err != nil {
+		t.Errorf("error when getting valid course")
+	}
+
+	_, err = GetCourseByNum("", 225)
+
+	if err == nil {
+		t.Errorf("should not be allowed to get empty subject")
+	}
+
+	_, err = GetCourseByNum("CS", -3)
+
+	if err == nil {
+		t.Errorf("should not be allowed to have non-three-digit number (too small)")
+	}
+
+	_, err = GetCourseByNum("CS", 1000)
+
+	if err == nil {
+		t.Errorf("should not be allowed to have non-three-digit number (too big)")
+	}
+}
